+++ packages/queries/src/Contracts/SortMethodFactoryInterface.php
@@ -10,20 +10,20 @@ namespace EDT\Querying\Contracts;
 interface SortMethodFactoryInterface
 {
     /**
-     * @param non-empty-list<non-empty-string> $properties
+     * @param non-empty-list<non-empty-string>|PropertyPathInterface $properties
      *
      * @return TSorting
      *
      * @throws PathException
      */
-    public function propertyAscending(array $properties): PathsBasedInterface;
+    public function propertyAscending($properties): PathsBasedInterface;
 
     /**
-     * @param non-empty-list<non-empty-string> $properties
+     * @param non-empty-list<non-empty-string>|PropertyPathInterface $properties
      *
      * @return TSorting
      *
      * @throws PathException
      */
-    public function propertyDescending(array $properties): PathsBasedInterface;
+    public function propertyDescending($properties): PathsBasedInterface;
 }